Add AVLTree, and demonstration block "__main__" works for all required Binary Tree, BST and AVL tree
However, I have still met some issues with the integrated unttest for all these trees:

When I have tried to call the unittest, all desired printings could not show up. So I will try to work on it. 

As for the final submit, I believe that the unit test is much advanced extension. So, if the deadline is the issue, I prefer not to including it in this .py file.