Update module9_lab.py
Edited the test_inorder, test_preorder, and test_postorder functions inside the TestBinaryTree class and test_insert_duplicate function in the TestAVLTree class to use contextlib to redirect stdout safely in the test functions. I also edited the assertIsNone in the test_remove_nonexistant function in the TestBST class to 
self.assertIsNotNone(self.bst.root, "Non-existent value should not be found"). Added a comment above the main block to explain how to run the demos for BinaryTree, BST, and AVLTree instead of the unit tests